Move heuristics into project root
